refactor business logic

+++ src/main/java/org/example/homework/PrintCurrency/CurrencyInput.java
@@ -2,6 +2,6 @@ package org.example.homework.PrintCurrency;
 
 public interface CurrencyInput {
 
-    long readAmount();
+    int readAmount();
 
 }
